Move graph rendering to a separate module
