+++ site/frontend/src/graph/render.ts
@@ -1,5 +1,5 @@
 import uPlot, {TypedArray} from "uplot";
-import {GraphData, GraphsSelector} from "../../graph/data";
+import {GraphData, GraphsSelector} from "./data";
 
 const commonCacheStateColors = {
   full: "#7cb5ec",
